+++ addons/im_livechat/static/src/components/channel_preview_view/channel_preview_view.js
@@ -1,12 +1,12 @@
 /** @odoo-module **/
 
-import { ThreadPreview } from '@mail/components/thread_preview/thread_preview';
+import { ThreadPreview } from '@mail/components/channel_preview_view/channel_preview_view';
 
 import { patch } from 'web.utils';
 
 const components = { ThreadPreview };
 
-patch(components.ThreadPreview.prototype, 'im_livechat/static/src/components/thread_preview/thread_preview.js', {
+patch(components.ThreadPreview.prototype, 'im_livechat/static/src/components/channel_preview_view/channel_preview_view.js', {
 
     //--------------------------------------------------------------------------
     // Public
