+++ addons/mail/static/src/components/channel_preview_view/channel_preview_view.xml
