+++ addons/mail/static/src/models/channel_preview_view.js
