+++ addons/rating/static/src/components/channel_preview_view/channel_preview_view.scss
