+++ addons/test_mail_full/static/tests/qunit_suite_tests/channel_preview_view_tests.js
@@ -3,7 +3,7 @@
 import { afterNextRender, start, startServer } from '@mail/../tests/helpers/test_utils';
 
 QUnit.module('test_mail_full', {}, function () {
-QUnit.module('thread_preview_tests.js');
+QUnit.module('channel_preview_view_tests.js');
 
 QUnit.test('rating value displayed on the thread preview', async function (assert) {
     assert.expect(4);
